+++ app/models/bank.rb
@@ -3,4 +3,5 @@ class Bank < ApplicationRecord
   belongs_to :guild
 
   has_many :bank_cells, dependent: :destroy
+  has_many :bank_requests, dependent: :destroy
 end