+++ app/models/game_item.rb
@@ -5,4 +5,5 @@ class GameItem < ApplicationRecord
   belongs_to :game_item_subcategory, optional: true
 
   has_many :bank_cells, dependent: :destroy
+  has_many :bank_requests, dependent: :destroy
 end