+++ spec/factories/bank_requests.rb
@@ -1,0 +1,8 @@
+FactoryBot.define do
+  factory :bank_request do
+    requested_amount { 10 }
+    status { 0 }
+    association :bank
+    association :game_item
+  end
+end