+++ spec/models/bank_spec.rb
@@ -1,6 +1,7 @@
 RSpec.describe Bank, type: :model do
   it { should belong_to :guild }
   it { should have_many(:bank_cells).dependent(:destroy) }
+  it { should have_many(:bank_requests).dependent(:destroy) }
 
   it 'factory should be valid' do
     bank = build :bank