+++ spec/models/game_item_spec.rb
@@ -3,6 +3,7 @@ RSpec.describe GameItem, type: :model do
   it { should belong_to(:game_item_category).optional }
   it { should belong_to(:game_item_subcategory).optional }
   it { should have_many(:bank_cells).dependent(:destroy) }
+  it { should have_many(:bank_requests).dependent(:destroy) }
 
   it 'factory should be valid' do
     game_item = build :game_item