+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - GameItem models
 - Parser for bank data
 - render bank cells
+- BankRequest model
 
 ## [1.0.1] - 2019-12-11
 ### Added